+++ StaffView/StaffView.Builder.swift
@@ -21,9 +21,7 @@ public enum LedgerLineDirection: Double {
 extension StaffView {
     
     public class Builder {
-        
-        
-        
+
         // Intput
         private let clef: Clef
         private let configuration: StaffConfiguration
@@ -31,13 +29,16 @@ extension StaffView {
         // Intermediate representation
         private var staffLines = LinesSegmentCollection()
         private var ledgerLines: [Double: [LedgerLineDirection: Int]] = [:]
-        private var points: [Double: [PointView]] = [:]
-        
+        private var points: [PointView] = []
+
+        /// Creates a `StaffView.Builder` ready to represent the given `clef`, with the given 
+        /// `configuration`.
         public init(clef: Clef, configuration: StaffConfiguration) {
             self.clef = clef
             self.configuration = configuration
         }
         
+        /// Creates an internal representation of the given `model` with the given `configuration`.
         public init(model: StaffModel, configuration: StaffConfiguration) {
             
             self.clef = model.verticalAxis
@@ -55,7 +56,7 @@ extension StaffView {
                         clef: model.verticalAxis,
                         staffSlotHeight: configuration.staffSlotHeight
                     )
-                    add(pointView, at: position)
+                    self.points.append(pointView)
                 }
             }
             
@@ -63,22 +64,32 @@ extension StaffView {
             stopLines(at: model.points.keys.max()! + 100)
         }
         
-        public func add(_ point: PointView, at position: Double) {
-            points.safelyAppend(point, toArrayWith: position)
-        }
-        
+        /// Starts staff lines at the given position.
         public func startLines(at x: Double) {
             staffLines.startLines(at: x)
         }
         
+        /// Stops staff lines at the given position.
         public func stopLines(at x: Double) {
             staffLines.stopLines(at: x)
         }
         
+        /// Adds the amount ledger lines `above` and `below` at the given `position`.
         public func addLedgerLines(at position: Double, above: Int, below: Int) {
             ledgerLines.ensureValue(for: position)
             ledgerLines[position]![.below] = below
             ledgerLines[position]![.above] = above
+        }
+        
+        /// Creates a `StaffView` with the
+        public func build() -> StaffView {
+            let clef = makeClef(with: configuration)
+            let lines = StaffLinesCollection(
+                staffLines: staffLines,
+                ledgerLines: ledgerLines,
+                configuration: configuration
+            )
+            return StaffView(clef: clef, lines: lines, points: points)
         }
         
         private func makeClef(with configuration: StaffConfiguration) -> StaffClefView {
@@ -95,88 +106,6 @@ extension StaffView {
             )
             
             return StaffClefView.makeClef(clef.kind, at: position, with: clefConfig)
-        }
-        
-//        private func staffLines(configuration: StaffConfiguration) -> StyledPath {
-//            
-//            let staffSlotHeight = configuration.staffSlotHeight
-//            
-//            let path = Path(
-//                staffLines.flatMap { segment in
-//                    return (0..<5).map { lineNumber in
-//                        let altitude = Double(lineNumber) * staffSlotHeight * 2
-//                        let left = segment.start
-//                        let right = segment.stop
-//                        return BezierCurve(
-//                            start: Point(x: left, y: altitude),
-//                            end: Point(x: right, y: altitude)
-//                        )
-//                    }
-//                }
-//            )
-//            
-//            let styling = Styling(
-//                stroke: Stroke(width: configuration.lineWidth, color: configuration.linesColor)
-//            )
-//            
-//            return StyledPath(frame: .zero, path: path, styling: styling)
-//        }
-//        
-//        private func ledgerLines(configuration: StaffConfiguration) -> StyledPath {
-//            
-//            let staffSlotHeight = configuration.staffSlotHeight
-//            let length = configuration.ledgerLineLength
-//            var curves: [BezierCurve] = []
-//            
-//            for (x, amountByDirection) in ledgerLines {
-//                for (direction, amount) in amountByDirection {
-//                    let left = x - 0.5 * length
-//                    let right = x + 0.5 * length
-//                    let refY = direction == .above ? -2 * staffSlotHeight : 10 * staffSlotHeight
-//                    
-//                    for number in 0..<amount {
-//                        let altitude = Double(number) * direction.rawValue * 2 * staffSlotHeight + refY
-//                        let curve = BezierCurve(
-//                            start: Point(x: left, y: altitude),
-//                            end: Point(x: right, y: altitude)
-//                        )
-//                        curves.append(curve)
-//                    }
-//                }
-//            }
-//            
-//            let path = Path(curves)
-//            
-//            let styling = Styling(
-//                stroke: Stroke(
-//                    width: configuration.ledgerLineWidth,
-//                    color: configuration.linesColor
-//                )
-//            )
-//            
-//            return StyledPath(frame: .zero, path: path, styling: styling)
-//        }
-//        
-//        private func makeLines(configuration: StaffConfiguration) -> StyledPath.Composite {
-//            
-//            let group = StyledPath.Group("lines")
-//            
-//            let paths = [
-//                staffLines(configuration: configuration),
-//                ledgerLines(configuration: configuration)
-//            ]
-//            
-//            return .branch(group, paths.map { .leaf($0) })
-//        }
-        
-        public func build() -> StaffView {
-            let clef = makeClef(with: configuration)
-            let lines = StaffLinesCollection(
-                staffLines: staffLines,
-                ledgerLines: ledgerLines,
-                configuration: configuration
-            )
-            return StaffView(clef: clef, lines: lines, points: points)
         }
     }
 }
